+++ src/RestrictedPython/compile.py
@@ -1,4 +1,5 @@
 from collections import namedtuple
+from RestrictedPython._compat import IS_CPYTHON
 from RestrictedPython._compat import IS_PY2
 from RestrictedPython.transformer import RestrictingNodeTransformer
 
@@ -11,6 +12,10 @@ CompileResult = namedtuple(
 syntax_error_template = (
     'Line {lineno}: {type}: {msg} in on statement: {statement}')
 
+NOT_CPYTHON_WARNING = (
+    'You are using on a not supported Python implementation.'
+    ' Use CPython to prevent security issues when using RestrictedPython!')
+
 
 def _compile_restricted_mode(
         source,
@@ -19,9 +24,13 @@ def _compile_restricted_mode(
         flags=0,
         dont_inherit=False,
         policy=RestrictingNodeTransformer):
+
+    if not IS_CPYTHON:
+        warnings.warn(NOT_CPYTHON_WARNING, RuntimeWarning)
+
     byte_code = None
-    errors = []
-    warnings = []
+    collected_errors = []
+    collected_warnings = []
     used_names = {}
     if policy is None:
         # Unrestricted Source Checks
@@ -43,24 +52,30 @@ def _compile_restricted_mode(
             try:
                 c_ast = ast.parse(source, filename, mode)
             except (TypeError, ValueError) as e:
-                errors.append(str(e))
+                collected_errors.append(str(e))
             except SyntaxError as v:
-                errors.append(syntax_error_template.format(
+                collected_errors.append(syntax_error_template.format(
                     lineno=v.lineno,
                     type=v.__class__.__name__,
                     msg=v.msg,
                     statement=v.text.strip()
                 ))
         if c_ast:
-            policy(errors, warnings, used_names).visit(c_ast)
-            if not errors:
+            policy_instance = policy(
+                collected_errors, collected_warnings, used_names)
+            policy_instance.visit(c_ast)
+            if not collected_errors:
                 byte_code = compile(c_ast, filename, mode=mode  # ,
                                     # flags=flags,
                                     # dont_inherit=dont_inherit
                                     )
     else:
         raise TypeError('Unallowed policy provided for RestrictedPython')
-    return CompileResult(byte_code, tuple(errors), warnings, used_names)
+    return CompileResult(
+        byte_code,
+        tuple(collected_errors),
+        collected_warnings,
+        used_names)
 
 
 def compile_restricted_exec(